userserver/src/requirements.txt
mysql-connector
requests
flask
prometheus_flask_exporter